+++ drivers/net/ppp_async.c
@@ -342,7 +342,7 @@ ppp_asynctty_poll(struct tty_struct *tty, struct file *file, poll_table *wait)
 /* May sleep, don't call from interrupt level or with interrupts disabled */
 static void
 ppp_asynctty_receive(struct tty_struct *tty, const unsigned char *buf,
-		  const char *cflags, int count)
+		     const char *cflags, size_t count)
 {
 	struct asyncppp *ap = ap_get(tty);
 	unsigned long flags;
