+++ sound/soc/codecs/cx20442.c
@@ -274,7 +274,7 @@ static int v253_hangup(struct tty_struct *tty)
 
 /* Line discipline .receive_buf() */
 static void v253_receive(struct tty_struct *tty, const unsigned char *cp,
-		const char *fp, int count)
+			 const char *fp, size_t count)
 {
 	struct snd_soc_codec *codec = tty->disc_data;
